dammit rog can't you do anything right

+++ build/ProcessorBot.js
@@ -103,11 +103,13 @@ class ProcessorBot {
             for (const guild of this.client.guilds.cache.values()) {
                 // @ts-ignore
                 const existingcommands = yield this.client.api.applications(this.client.user.id).guilds(guild.id).commands.get();
-                for (const command of existingcommands) {
-                    // console.log(command);
-                    // @ts-ignore
-                    let currDeletePromise = this.client.api.applications(this.client.user.id).guilds(guild.id).commands(command.id).delete();
-                    allDeletePromises.push(currDeletePromise);
+                if (existingcommands) {
+                    for (const command of existingcommands) {
+                        // console.log(command);
+                        // @ts-ignore
+                        let currDeletePromise = this.client.api.applications(this.client.user.id).guilds(guild.id).commands(command.id).delete();
+                        allDeletePromises.push(currDeletePromise);
+                    }
                 }
             }
             yield Promise.all(allDeletePromises);
